refactor(account): B2B-2146 Add individual methods per formField type
We were sending an integer to the API to retrieve the different type of
formFields, this PR introduces 3 different methods, this simplifies the
code to retrieve methods and better documents what's being retrieved.

Simplify the fetch/mapping of form fields in the registration page.

Bonus: Straighten some of the types around countries.